reboot on speciffic hook exit status
When a hook exits with error code 1001, juju will requeue the hook,
issue a reboot and Stop(). This will cause juju to resume processing the hook
after reboot. It will be the charm author's job to make sure the hook checks
if another reboot is needed or not, or if its fine to continue with the normal
hook operation.